fix: resolve all compiler warnings and clippy lints across workspace
- Fix unused imports in minimal_telemetry_test.rs
- Add #![allow(dead_code)] to test helper modules
- Collapse nested if-let chains using && in standalone/marker, region, midi, item, automation
- Replace len() > 0 checks with !is_empty() in test files
- Replace manual strip_prefix patterns with .strip_prefix() in fx_chain.rs
- Prefix unused variables with _ in test files
- Fix unused Result values in daw-reaper/src/item.rs
- Apply fixes across daw-proto, daw-reaper, daw-standalone, dawfile-reaper, daw-ui, audio-controls

+++ crates/daw-reaper/src/marker.rs
@@ -55,7 +55,7 @@ impl MarkerService for ReaperMarker {
                         // region_end_position is None for markers, Some for regions
                         if info.region_end_position.is_none() {
                             markers.push(Marker {
-                                id: Some(info.id.get() as u32),
+                                id: Some(info.id.get()),
                                 position: Position::from_time(TimePosition::from_seconds(
                                     info.position.get(),
                                 )),
